[FIX] Fix issue that restrict to 10 variants
Elastic client limit by 10 the number of record return
In case of searching a product we when all variants without limit
This quick fix set to 10000 the limit.
If we have more then 10000 variants we will have an issue.
But having 10000 variants should never occured and maybe we should
manage them in a different way.
So let's see how to fix this issue the day we will have it.
One possible thing is to add a optionnal config and keeping this
default value.